fix:功能优化

+++ ubold-orm/src/main/java/com/ubold/admin/pager/Pager.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Data
 public class Pager<T> implements Serializable{
 	private static final long serialVersionUID = 1L;
-	private Integer pageSize;
-	private Integer pageNo;
-	private Integer totalElements;
+	private int pageSize;
+	private int pageNo;
+	private Long totalElements;
 	private List<T>	content;
 }
